feat: Move all components except App to comp dir (#157)
