+++ src/components/Conversation.tsx
@@ -1,11 +1,11 @@
 import styles from './ChatView.module.css';
 import { Content } from './Content';
 import { StreamingText } from './StreamingText';
-import { messageStore, progressStore, toolCallStreamStore } from './store';
+import { messageStore, progressStore, toolCallStreamStore } from '../store';
 import type { ChatCompletionMessageParam } from 'openai/resources/index';
 import { memo } from 'react';
-import { useTheme } from './theme/useTheme';
-import { useAppContext } from './context/useAppContext';
+import { useTheme } from '../theme/useTheme';
+import { useAppContext } from '../context/useAppContext';
 
 export interface ConversationProps {
   messages: ChatCompletionMessageParam[];
